+++ mobile/app/_layout.tsx
@@ -685,6 +685,10 @@ function RootNavigator() {
         options={{ ...modalOptions, headerShown: false, title: t('eventRegistration:guests.title') }}
       />
       <Stack.Screen
+        name="(modals)/event-registration-invitations"
+        options={{ ...modalOptions, headerShown: false, title: t('eventRegistration:invitations.builder_title') }}
+      />
+      <Stack.Screen
         name="(modals)/event-templates"
         options={{ ...modalOptions, headerShown: false, title: t('event_templates:templates.mobile.title') }}
       />